moved getDocuments into helper file. featureCollectionTitle can be sublayerTitle or layerTitle.
