+++ src/js/controllers/mapController.ts
@@ -1073,45 +1073,6 @@ export class MapController {
     }
   }
 
-  async getDocuments(
-    urlProperties: URLProperties
-  ): Promise<Array<Attachment> | null> {
-    const { sublayerID, specificFeatureID, layerID } = urlProperties;
-    let endPoint = '';
-    const layer = this._map?.findLayerById(layerID);
-
-    if ((layer as any).sublayers.length && specificFeatureID) {
-      const sublayer = (layer as any).sublayers.items.filter(
-        (s: Sublayer) => s.id === sublayerID
-      );
-
-      endPoint = `${sublayer[0].url}/${specificFeatureID}/attachments?f=pjson`;
-    } else {
-      return null;
-    }
-
-    const attachments = await fetch(endPoint)
-      .then(response => response.json())
-      .then((results: { attachmentInfos: Array<Attachment> }) => {
-        const { attachmentInfos } = results;
-
-        return attachmentInfos.map((attachment: Attachment) => {
-          attachment.url = endPoint.replace(
-            'attachments?f=pjson',
-            `attachments/${results.attachmentInfos[0].id}`
-          );
-          return attachment;
-        });
-      })
-      .catch(e => console.log('error in getDocuments()', e));
-
-    if (attachments && attachments.length) {
-      return attachments;
-    } else {
-      return null;
-    }
-  }
-
   processGeojson(esriJson: Array<FeatureResult>): void {
     if (this._map) {
       setNewGraphic({
